Edited classes.py again.
Change 1: Edited _Connection

When we initialize the file, we have to go through each player's passes one by one. So there is a risk that we end up changing the connection twice if two players interact with each other. What i've done here is that in 'tweak_stats' in our initialization i've set which player is player1 and player2 so this function checks which one is player1 or player 2 and changes the stats accordingly. (player1_avg...)

Then I added finalzie stats to just finalize the max of the data, will add more depending on how many more stats we want to create.

Change 2: Graph class

Added add_player: Nothing much to say here

Added _Connections instance, makes sure that we keep track of connections and KEEP THEM UNIQUE

after that I added add_connection which checks if the connection already exists or not. If it doesnt, it creates the new conenction and sets the PLAYER1 PLAYER CLASS to be player 1 and the PLAYER2 string (name) to be PLAYER2.

Then regardless if the connection exists or not, i call 'tweak stats' from the connection stats to tweak the stats.

Please do read all of this and DM if you have questions.